moved: model classes to data module
